Update Swap_nodes.cpp
The basic idea is to create a tree first, and then get all the nodes present on the given level in a vector. Now go through all the elements in the vector and then swap them and get the inorder of the tree are this operation and store this in the result vector. Next, go through all the multiples of the given level and similarly get the nodes and then swap them and update the results.